updated injectable usage in services to be provided in root; patched standard guideines calls; patched blog calls
