+++ src/app/auth/register/components/recaptcha/recaptcha-validator.service.ts
@@ -1,10 +1,12 @@
-import {map} from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 import { USER_ROUTES } from '@env/route';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { AbstractControl } from '@angular/forms';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class RecaptchaValidator {
 
     constructor(private http: HttpClient) {
@@ -13,12 +15,12 @@ export class RecaptchaValidator {
     validateToken(token: string) {
         return (_: AbstractControl) => {
             return this.http.get(USER_ROUTES.VALIDATE_CAPTCHA(), { params: { token } }).pipe(
-            map((res: any) => {
-                if (!res.success) {
-                    return { tokenInvalid: true };
-                }
-                return null;
-            }));
+                map((res: any) => {
+                    if (!res.success) {
+                        return { tokenInvalid: true };
+                    }
+                    return null;
+                }));
         };
     }
 }
