+++ src/app/core/client-module/mappings-filter.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@angular/core';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class MappingsFilterService {
   // to use this service, provide this in the parent component of the browse-by-mappings component
   constructor() { }
