+++ src/app/core/interceptor/httpconfig.interceptor.ts
@@ -10,7 +10,9 @@ import { Observable } from 'rxjs';
 import { CookieService } from 'ngx-cookie';
 import { environment } from '@env/environment';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class HttpConfigInterceptor implements HttpInterceptor {
   private token: string;
   constructor(private cookie: CookieService) {
@@ -20,9 +22,10 @@ export class HttpConfigInterceptor implements HttpInterceptor {
     request: HttpRequest<any>,
     next: HttpHandler
   ): Observable<HttpEvent<any>> {
+    console.log('intercepted request ... ');
     let headers = request.headers.set('Content-Type', 'application/json');
     if (!request.url.includes(environment.cardOrganizationUrl)) {
-    headers = headers.append('Authorization', `Bearer ${this.token}`);
+      headers = headers.append('Authorization', `Bearer ${this.token}`);
     }
 
     request = request.clone({
