+++ src/app/core/learning-object-module/custom-url-serliazer.ts
@@ -1,7 +1,9 @@
-import {UrlSerializer, UrlTree, DefaultUrlSerializer} from '@angular/router';
+import { UrlSerializer, UrlTree, DefaultUrlSerializer } from '@angular/router';
 import { Injectable } from '@angular/core';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class CustomUrlSerializer implements UrlSerializer {
     parse(url: any): UrlTree {
         const dus = new DefaultUrlSerializer();
