+++ src/app/core/learning-object-module/editor.service.ts
@@ -4,7 +4,9 @@ import { HttpClient } from '@angular/common/http';
 import { catchError } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class EditorService {
   constructor(private http: HttpClient) { }
