+++ src/app/core/metric-module/metric.service.ts
@@ -4,7 +4,9 @@ import { catchError, retry } from 'rxjs/operators';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { throwError } from 'rxjs';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class MetricService {
 
   constructor(private http: HttpClient) { }
