+++ src/app/cube/browse/components/filter/filter.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectorRef, Component, OnInit, Output, EventEmitter, OnDestroy, Input } from '@angular/core';
 import { LearningObject } from '@entity';
 import { CollectionService } from 'app/core/collection-module/collections.service';
-import { GuidelineService } from 'app/core/standard-guidelines-module/guideline.service';
+import { GuidelineService } from 'app/core/standard-guidelines-module/standard-guidelines.service';
 import { RelevancyService } from 'app/core/learning-object-module/relevancy.service';
 import { Observable, Subject } from 'rxjs';
 import { debounceTime, takeUntil } from 'rxjs/operators';
