+++ src/app/onion/core/suggestion.service.ts
@@ -5,7 +5,9 @@ import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http
 import { environment } from '@env/environment';
 import * as querystring from 'querystring';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class SuggestionService {
   suggestion = new Subject<{}[]>();
   mappedStandards = [];
