+++ src/app/onion/relevancy-builder/components/standard-outcomes/standard-outcomes.component.ts
@@ -8,7 +8,7 @@ import {
   EventEmitter,
 } from '@angular/core';
 import { Guideline } from '@entity';
-import { GuidelineService } from 'app/core/standard-guidelines-module/guideline.service';
+import { GuidelineService } from 'app/core/standard-guidelines-module/standard-guidelines.service';
 import { BuilderStore } from '../../builder-store.service';
 import { BehaviorSubject, Subject, Subscription } from 'rxjs';
 import { takeUntil, debounceTime, filter, map } from 'rxjs/operators';
